feat(ui): add client page for select connector
