+++ src/components/Catalog/Catalog.tsx
@@ -1,41 +1,13 @@
 'use client';
 import { Catalog, ITile } from '@kaoto-next/ui';
-import { useCallback } from 'react';
-import {
-  PageSection,
-  Split,
-  SplitItem,
-  Text,
-  TextContent,
-  Title,
-} from '@patternfly/react-core';
 
 interface ICatalogContainer {
+  onTileClick: (_tile: ITile) => void;
   tiles: Record<string, ITile[]>;
 }
 
-const CatalogContainer = ({ tiles }: ICatalogContainer) => {
-  const onTileClick = useCallback((tile: ITile) => {
-    console.log('Tile clicked', tile);
-  }, []);
-
-  return (
-    <>
-      <PageSection variant={'light'}>
-        <Split>
-          <SplitItem isFilled>
-            <TextContent>
-              <Title headingLevel="h1">Create Flow</Title>
-              <Text>Lorem ipsum dolorum</Text>
-            </TextContent>
-          </SplitItem>
-        </Split>
-      </PageSection>
-      <PageSection>
-        <Catalog tiles={tiles} onTileClick={onTileClick} />
-      </PageSection>
-    </>
-  );
+const CatalogContainer = ({ onTileClick, tiles }: ICatalogContainer) => {
+  return <Catalog tiles={tiles} onTileClick={onTileClick} />;
 };
 
 export default CatalogContainer;
